feat: add dedicated cart page with header navigation
