+++ src/App.tsx
@@ -1,3 +1,4 @@
+import { useState } from "react";
 import { Toaster } from "@/components/ui/toaster";
 import { Toaster as Sonner } from "@/components/ui/sonner";
 import { TooltipProvider } from "@/components/ui/tooltip";
@@ -5,23 +6,91 @@ import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Index from "./pages/Index";
 import NotFound from "./pages/NotFound";
+import Cart from "./pages/Cart";
+import { Header } from "@/components/Header";
+import { PRODUCTS, Product } from "@/lib/products";
+import { useToast } from "@/hooks/use-toast";
 
 const queryClient = new QueryClient();
 
-const App = () => (
-  <QueryClientProvider client={queryClient}>
-    <TooltipProvider>
-      <Toaster />
-      <Sonner />
-      <BrowserRouter>
-        <Routes>
-          <Route path="/" element={<Index />} />
-          {/* ADD ALL CUSTOM ROUTES ABOVE THE CATCH-ALL "*" ROUTE */}
-          <Route path="*" element={<NotFound />} />
-        </Routes>
-      </BrowserRouter>
-    </TooltipProvider>
-  </QueryClientProvider>
-);
+interface CartItem extends Product {
+  quantity: number;
+}
+
+const App = () => {
+  const [cart, setCart] = useState<Record<string, number>>({});
+  const [orderPlaced, setOrderPlaced] = useState<string | null>(null);
+  const { toast } = useToast();
+
+  const addToCart = (productId: string) => {
+    setCart(prev => ({
+      ...prev,
+      [productId]: (prev[productId] || 0) + 1
+    }));
+    toast({
+      title: "Added to cart",
+      description: "Item has been added to your order",
+    });
+  };
+
+  const updateQuantity = (productId: string, quantity: number) => {
+    setCart(prev => {
+      if (quantity === 0) {
+        const newCart = { ...prev };
+        delete newCart[productId];
+        return newCart;
+      }
+      return { ...prev, [productId]: quantity };
+    });
+  };
+
+  const getCartItems = (): CartItem[] => {
+    return Object.entries(cart)
+      .map(([productId, quantity]) => {
+        const product = PRODUCTS.find(p => p.id === productId);
+        return product ? { ...product, quantity } : null;
+      })
+      .filter((item): item is CartItem => item !== null);
+  };
+
+  const handleCheckout = () => {
+    const orderNumber = `ORD-${Date.now().toString().slice(-6)}`;
+    console.log("Order placed:", {
+      orderNumber,
+      items: getCartItems(),
+      timestamp: new Date().toISOString()
+    });
+    setOrderPlaced(orderNumber);
+    setCart({});
+    toast({
+      title: "Order placed successfully!",
+      description: `Order ${orderNumber} has been submitted for review.`,
+    });
+  };
+
+  const resetToShopping = () => {
+    setOrderPlaced(null);
+  };
+
+  const totalItems = getCartItems().reduce((sum, item) => sum + item.quantity, 0);
+
+  return (
+    <QueryClientProvider client={queryClient}>
+      <TooltipProvider>
+        <Toaster />
+        <Sonner />
+        <BrowserRouter>
+          <Header cartItemCount={totalItems} />
+          <Routes>
+            <Route path="/" element={<Index cart={cart} addToCart={addToCart} updateQuantity={updateQuantity} />} />
+            <Route path="/cart" element={<Cart items={getCartItems()} onCheckout={handleCheckout} orderPlaced={orderPlaced} resetToShopping={resetToShopping} />} />
+            {/* ADD ALL CUSTOM ROUTES ABOVE THE CATCH-ALL "*" ROUTE */}
+            <Route path="*" element={<NotFound />} />
+          </Routes>
+        </BrowserRouter>
+      </TooltipProvider>
+    </QueryClientProvider>
+  );
+};
 
 export default App;
